FE code quality

+++ backend/controllers/PostController.ts
@@ -1,6 +1,8 @@
-import { FileService } from "../service/FileService";
 import { PostService } from "../service/PostService";
 import { Request, Response, NextFunction } from "express";
+import MissingCoordinateException from "../errors/customError";
+import { PostModel } from "../model/PostModel";
+
 interface ImageData {
     fileData: Buffer;
     fileType: string;
@@ -8,11 +10,9 @@ interface ImageData {
 
 export class PostController {
     private postService: PostService;
-    private fileService: FileService;
 
     constructor() {
         this.postService = new PostService();
-        this.fileService = new FileService();
     }
 
     createPost = async (req: Request, res: Response, next: NextFunction) => {
@@ -26,42 +26,65 @@ export class PostController {
         res.json(await this.postService.createPost(post))
     }
 
-    getPosts = async (req: Request, res: Response, next: NextFunction) => {
-        const { start_lat, end_lat, start_long, end_long, isPrivate } = req.query;
-        const user_id = undefined // TODO: Find user id right here
-        // Parse the coordinates if present
-        const parsedStartLat = start_lat ? parseFloat(start_lat as string) : undefined;
-        const parsedEndLat = end_lat ? parseFloat(end_lat as string) : undefined;
-        const parsedStartLong = start_long ? parseFloat(start_long as string) : undefined;
-        const parsedEndLong = end_long ? parseFloat(end_long as string) : undefined;
-    
-        // Parse the "isPrivate" parameter as a boolean
-        const isPostPrivate = isPrivate === 'true'; // "true" or "false" as strings from query param
-    
-        // Optionally log or handle missing params
-        if (!parsedStartLat || !parsedEndLat || !parsedStartLong || !parsedEndLong) {
-            return res.status(400).json({ message: "Missing required query parameters" });
+    getPublicPost = async (req: Request, res: Response, next: NextFunction) => {
+        try{
+            const {parsedStartLat, parsedEndLat, parsedStartLong, parsedEndLong } =  parseLocationParam(req)
+        
+            // Call service with the query params
+            res.json(await this.postService.getPublicPost(
+                parsedStartLat, 
+                parsedEndLat, 
+                parsedStartLong, 
+                parsedEndLong,
+            ));
         }
-    
-        // Call your service with the query params
-        res.json(await this.postService.getPosts(
-            user_id as any,
-            parsedStartLat, 
-            parsedEndLat, 
-            parsedStartLong, 
-            parsedEndLong,
-            isPostPrivate 
-        ));
+        catch(error){
+            if(error instanceof MissingCoordinateException){
+                console.log("User Provided Invalid coordinate: ", error)
+                return res.status(400).json({ message: "Incomplete coordinate" });
+            }
+            else{
+                console.log("Error: ", error);
+                return res.status(500).json({ message: "Internal Server Error" });
+
+            }
+        }
     };
+
+ 
+    getAuthenticatedUserPost = async(req: Request, res: Response, next: NextFunction) => {
+        try{
+            const {userPostOnly, start_lat, end_lat, start_long, end_long} = req.query;
+
+            // Parse the coordinates if present
+            const parsedStartLat = start_lat ? parseFloat(start_lat as string) : undefined;
+            const parsedEndLat = end_lat ? parseFloat(end_lat as string) : undefined;
+            const parsedStartLong = start_long ? parseFloat(start_long as string) : undefined;
+            const parsedEndLong = end_long ? parseFloat(end_long as string) : undefined;
+
+            const parsedUserPostOnly = userPostOnly == "true"
+
+            const userId = (req as any).user.id
+            res.json(await this.postService.getUserPost(userId, parsedUserPostOnly, parsedStartLat,
+                                                        parsedEndLat, parsedStartLong,
+                                                        parsedEndLong))
+        }
+
+        catch(err){
+            console.log("ERROR: ", err)
+            return res.status(500).json({message: "Internal Server Error"})
+        }
+
+
+    }
 
     getPostById = async (req: Request, res: Response, next: NextFunction) => {
         const postId = req.params.id
-        const [fileData, postData] = await Promise.all([
-            this.fileService.getFileInPost(postId),
+        const [postData] = await Promise.all([
             this.postService.getPostById(postId)
         ]);
           
-        res.json({ postData, fileId: fileData});
+        res.json({ postData });
     }
 
     updatePost = async (req: Request, res: Response, next: NextFunction) => {
@@ -73,6 +96,46 @@ export class PostController {
 
     deletePost = async (req: Request, res: Response, next: NextFunction) => {
         const postId = req.params.id
-        res.json(await this.postService.deletePost(postId))
+
+        // Ensure that the post really belongs to the user
+        const userId = (req as any).user.id
+        const post = await PostModel.findById(postId)
+
+        if(post?.userId !== userId){
+            res.status(401).send({message: "Unauthorized"})
+        }
+        else{
+            res.json(await this.postService.deletePost(postId))
+        }
     }
+
+    getEveryPost = async (req: Request, res: Response, next: NextFunction) => {    
+        res.json(await this.postService.getEveryPost());
+    }
+
+    getPostsAtLocation = async (req: Request, res: Response, next: NextFunction) => {
+        const coordinates = (req as any).query    
+        res.json(await this.postService.getPostsAtLocation(coordinates.lat, coordinates.long));
+    }
+}
+
+/**
+ * Parse the longitude and latitude information of a request param
+ */
+function parseLocationParam(req: Request){
+    const { start_lat, end_lat, start_long, end_long} = req.query;
+
+    // Parse the coordinates if present
+    const parsedStartLat = start_lat ? parseFloat(start_lat as string) : undefined;
+    const parsedEndLat = end_lat ? parseFloat(end_lat as string) : undefined;
+    const parsedStartLong = start_long ? parseFloat(start_long as string) : undefined;
+    const parsedEndLong = end_long ? parseFloat(end_long as string) : undefined;
+
+    return {
+        parsedStartLat: parsedStartLat,
+        parsedEndLat: parsedEndLat,
+        parsedStartLong: parsedStartLong,
+        parsedEndLong: parsedEndLong
+    }
+
 }
